add proper RSS support for a wiki style website
I have a tag called "evergeen" (though that might misleading? maybe
called it "complete") that signifies that it is not a stub, that it is
a complete thought.

another way is to:

- base it on just X amount of words: this means no special tag, but it
feels messy. I could a ton of words but it would be disjointed.

The rss feed should serve a purpose and give you something useful and
not disjointed stubs.